requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj_database_url
Django
django-autoslug
django-cors-headers
django-taggit
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
gunicorn
idna
oauthlib
pillow
psycopg2
psycopg2-binary
pycparser
PyJWT
python-decouple
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3
whitenoise
